telemetry-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
redis>=5.0.0
celery>=5.3.0
python-jose[cryptography]
psycopg2-binary